backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pymongo==4.6.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
pydantic==2.4.2
google-generativeai==0.3.1
pydantic-settings==2.0.3 